--- src/pages/LoginSignupPage/LoginPage.css
@@ -1,101 +1,0 @@
-.login-container {
-    border: 1px solid lightgray;
-    max-width: 900px;
-    height: 500px;
-    width: 100%;
-    background-color: #F0F0F0;
-    padding: 20px;
-    align-items: center;
-    display: flex;
-    justify-content: space-between;
-    border-radius: 20px;
-}
-
-.emoji-area {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    flex: 1;
-}
-
-.emoji-text-area {
-    text-align: center;
-    color: gray;
-}
-
-.emoji-text1,
-.emoji-text2 {
-    font-weight: bold;
-}
-
-.login-area {
-    flex: 2;
-}
-
-.login-box label {
-    font-weight: 700;
-}
-
-.login-box-header {
-    text-align: center;
-    margin: 30px;
-    font-weight: bold;
-    color: rgb(169, 170, 133);
-}
-
-.mb-3 {
-    width: 70%;
-    max-width: 400px;
-    max-height: 200px;
-    margin: 0 auto;
-}
-
-
-
-.mb-3 .form-control {
-    height: 50px;
-    font-size: 16px;
-}
-
-.display-center {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-
-.button-box {
-    width: 70%;
-    max-width: 400px;
-    margin: 20px auto 0 auto;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.login-button {
-    width: 100%;
-    height: 50px;
-    background-color: rgb(169, 170, 133)!important; 
-    /* !important 해야 기본 부트스트랩 css 위로 적용됌. */
-    color: #ffffff!important;
-    border: none!important;
-    font-size: 16px;
-    margin-bottom: 10px;
-}
-
-.login-button:hover {
-    background-color: rgb(143, 144, 102)!important;
-}
-
-.login-button:active {
-    background-color: rgb(120, 121, 85)!important;
-    /* 버튼 클릭 시 색상 */
-    transform: scale(0.98)!important;
-    /* 클릭 시 약간 눌리는 효과 (선택사항) */
-    outline: none!important;
-    /* 기본 스타일 제거 */
-}
-
