+++ src/pages/LoginSignupPage/LoginPage/LoginPage.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import './LoginPage.css';
+
+
+const LoginPage = () => {
+  return (
+    <div>
+     Login Page
+    </div>
+  );
+};
+
+export default LoginPage;
